% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sites_from_files.R
\name{create_connection_from_files}
\alias{create_connection_from_files}
\title{creates connection from RData files and sample information}
\usage{
create_connection_from_files(sampleInfo, sites_final_path,
  ref_genome = "hg18")
}
\arguments{
\item{sampleInfo}{file path to sample information(tab separated file)}

\item{sites_final_path}{vector of file paths to sites.final.RData}

\item{ref_genome}{name of genome, at present the same
genome is used for all the samples}
}
\value{
connection (represented as a list)
connection has: sites df, sample_sex df, sitesFromFiles members
}
\description{
creates connection from RData files and sample information
}
\note{
stop if sample names from RData do not present in sample information file

list has member sitesFromFiles that is TRUE

connection is used instead of DB connection

required columns are alias(synonym of setName, sampleNames) and gender
}

