% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random_site.R
\name{get_N_MRCs}
\alias{get_N_MRCs}
\title{generate random controls for sites}
\usage{
get_N_MRCs(sites_meta, reference_genome, number_mrcs_per_site = 3)
}
\arguments{
\item{sites_meta}{dataframe with columns: siteID, gender}

\item{reference_genome}{BS genome object. All sites have the same genome.}

\item{number_mrcs_per_site}{number of MRCs to generate for each site}
}
\value{
dataframe with columns: siteID, chr, strand, position
}
\description{
generate random controls for sites
}
\note{
siteID are the same as given by sites_meta df
}

