% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_site.R
\name{get_random_positions}
\alias{get_random_positions}
\title{for a given reference genome and gender generate random positions}
\usage{
get_random_positions(siteIDs, reference_genome, gender = "m",
  number_of_positions = 3, male_chr = c("chrY"))
}
\arguments{
\item{siteIDs}{vector of unique siteIDs for use as random seed}

\item{reference_genome}{BS object reference genome(@seealso get_reference_genome)}

\item{gender}{'m' or 'f'}

\item{number_of_positions}{total number of random positions to generate}

\item{male_chr}{list of male-specific chromosomes prefixes(only 1 prefix is allowed at present)}
}
\value{
dataframe with columns: 
     siteID(numeric), chr(character), strand(character), position(numeric)
}
\description{
for a given reference genome and gender generate random positions
}
