% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_site.R
\name{get_reference_genome}
\alias{get_reference_genome}
\title{return BS genome OBJECT(not name) for human readable UCSC format}
\usage{
get_reference_genome(reference_genome)
}
\description{
format examples are: hg18, hg19, hg38 for human
                     mm8, mm9 for mouse
}
\note{
stop if cannot find unique genome from installed BSgenome
}
\seealso{
getRefGenome
}
