% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getMRCs}
\alias{getMRCs}
\title{creates match random controls.}
\usage{
getMRCs(setName, numberOfMRCs = 3, conn = NULL)
}
\arguments{
\item{setName}{vector of sample names}

\item{numberOfMRCs}{how many controls for each site}

\item{conn}{connection: DB or File connection}
}
\description{
creates match random controls.
}

