% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getMultihitLengths}
\alias{getMultihitLengths}
\title{lengths distributions for multihits}
\usage{
getMultihitLengths(sampleName, conn = NULL)
}
\arguments{
\item{sampleName}{vector of sample names}

\item{conn}{connection: DB or File connection}
}
\value{
df with 3 cols: sampleName, multihitID, length
}
\description{
lengths distributions for multihits
}

