% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getMultihitPositions}
\alias{getMultihitPositions}
\title{multihits}
\usage{
getMultihitPositions(setName, conn = NULL)
}
\arguments{
\item{setName}{vector of sample names}

\item{conn}{connection: DB or File connection}
}
\description{
multihits
}

