% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getRefGenome}
\alias{getRefGenome}
\title{name of the reference genome used}
\usage{
getRefGenome(setName, conn = NULL)
}
\arguments{
\item{setName}{vector of sample names}

\item{conn}{connection: DB or File connection
return  df with 2 cols: sampleName and refGenome}
}
\description{
name of the reference genome used
}

