% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getSampleNamesLike}
\alias{getSampleNamesLike}
\title{find replicates}
\usage{
getSampleNamesLike(setName, conn = NULL)
}
\arguments{
\item{setName}{vector of sample names}

\item{conn}{connection: DB or File connection}
}
\description{
find replicates
}
\note{
only function that treats % as a wildcard rather than a literal
}

