% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{setNameExists}
\alias{setNameExists}
\title{do we have sample names for a given connection}
\usage{
setNameExists(sample_ref, conn)
}
\arguments{
\item{conn}{connection: DB or File connection}

\item{sample_ref:}{df with 2 cols: sampleName, refGenome}
}
\value{
vector of TRUE/FALSE
}
\description{
do we have sample names for a given connection
}

