% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getMRCs}
\alias{getMRCs}
\title{creates match random controls.}
\usage{
getMRCs(sample_ref, conn, numberOfMRCs = 3)
}
\arguments{
\item{numberOfMRCs}{how many controls for each site}

\item{sampe_ref}{df with 2 cols: sampleName, refGenome}

\item{connection:}{DB or File connection}
}
\value{
df with cols: siteID, position, strand, chr, sampleName, refGenome
}
\description{
creates match random controls.
}

