% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getMRCs}
\alias{getMRCs}
\title{creates match random controls.}
\usage{
getMRCs(sample_ref, conn, numberOfMRCs = 3)
}
\arguments{
\item{sample_ref}{dataframe with 2 cols: sampleName, refGenome}

\item{conn}{connection to database}

\item{numberOfMRCs}{how many controls for each site}
}
\value{
df with cols: siteID, position, strand, chr, sampleName, refGenome
}
\description{
creates match random controls.
}
