% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getMultihitLengths}
\alias{getMultihitLengths}
\title{lengths distributions for multihits}
\usage{
getMultihitLengths(sample_ref, conn)
}
\arguments{
\item{conn}{DB connection}

\item{sampleName}{vector of sample names}
}
\value{
df with 3 cols: sampleName, refGenome, multihitID, length
}
\description{
lengths distributions for multihits
}

