% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getUniquePCRbreaks}
\alias{getUniquePCRbreaks}
\title{breakpoints}
\usage{
getUniquePCRbreaks(sample_ref, conn)
}
\arguments{
\item{sample_ref}{dataframe with 2 cols: sampleName, refGenome}

\item{conn}{connection to database}
}
\description{
breakpoints
}

