% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getUniquePCRbreaks}
\alias{getUniquePCRbreaks}
\title{breakpoints}
\usage{
getUniquePCRbreaks(setName, conn = NULL)
}
\arguments{
\item{setName}{vector of sample names}

\item{conn}{connection: DB or File connection}
}
\description{
breakpoints
}

