% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getUniqueSiteCounts}
\alias{getUniqueSiteCounts}
\title{unique counts for integration sites for a given sample(with fixed genome)}
\usage{
getUniqueSiteCounts(sample_ref, conn)
}
\arguments{
\item{sample_ref}{dataframe with 2 cols: sampleName, refGenome}

\item{conn}{connection to database}
}
\description{
unique counts for integration sites for a given sample(with fixed genome)
}
