% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getUniqueSiteReadCounts}
\alias{getUniqueSiteReadCounts}
\title{counts}
\usage{
getUniqueSiteReadCounts(setName, conn = NULL)
}
\arguments{
\item{setName}{vector of sample names}

\item{conn}{connection: DB or File connection}
}
\description{
counts
}

