% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getUniqueSiteReadCounts}
\alias{getUniqueSiteReadCounts}
\title{counts}
\usage{
getUniqueSiteReadCounts(sample_ref, conn)
}
\arguments{
\item{sample_ref}{dataframe with 2 cols: sampleName, refGenome}

\item{conn}{connection to database}
}
\description{
counts
}
