% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getUniqueSiteReadCounts}
\alias{getUniqueSiteReadCounts}
\title{counts}
\usage{
getUniqueSiteReadCounts(sample_ref, conn)
}
\arguments{
\item{conn}{DB connection}

\item{setName}{vector of sample names}
}
\description{
counts
}

