% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{getUniqueSites}
\alias{getUniqueSites}
\title{for a given sample names get sites.}
\usage{
getUniqueSites(setName, conn = NULL)
}
\arguments{
\item{setName}{vector of sample names}
}
\description{
for a given sample names get sites.
}

