% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intSiteRetriever.R
\name{setNameExists}
\alias{setNameExists}
\title{do we have samples in database}
\usage{
setNameExists(sample_ref, conn)
}
\arguments{
\item{sample_ref}{dataframe with 2 cols: sampleName, refGenome}

\item{conn}{connection to database}
}
\value{
vector of TRUE/FALSE for each row in sample_ref df
}
\description{
do we have samples in database
}
